+++ CHANGES.md
@@ -3,6 +3,7 @@ Version 0.3 in progress
 
 - repo_activity command accept repository parameter to filter which repository should be included in the report
 - repo_activity command accept additional order_by fields
+- find_events can by filtered by actor login
 
 
 Version 0.2 2012-12-20